requirements.txt
Django==3.2.5
djangorestframework==3.12.4
drf-spectacular==0.17.2
djangorestframework-camel-case==1.2.0
django-extensions==3.1.3
django-rest-auth[with_social]==0.9.5
django-filter==2.4.0
drf-exceptions-hog==0.0.5
psycopg2-binary==2.9.1
Pillow==8.3.1
whitenoise==5.3.0